launcher v0.4.4
deleted fomod
