+++ skymp5-launcher/src/renderer/index.html
@@ -162,8 +162,8 @@
             <label class="setting-label">Modlist Installation</label>
             <div class="vortex-status-row" id="mo2-status-row">
               <span class="vortex-status-dot" id="mo2-status-dot"></span>
-              <span class="vortex-status-text" id="mo2-status-text">Checking…</span>
               <button class="btn-install" id="btn-install-mo2">Install Modpack via MO2</button>
+              <span class="vortex-status-text" id="mo2-status-text">Checking…</span>
             </div>
             <div class="install-status" id="install-status-mo2"></div>
           </div>
